fix the user / post relationship

+++ controllers/feed.js
@@ -3,7 +3,8 @@ const path = require('path');
 
 const { validationResult } = require('express-validator');
 
-const Post = require('../models/Post');
+const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
 	const currentPage = req.query.page || 1;
@@ -48,21 +49,29 @@ exports.createPost = (req, res, next) => {
 		const imageUrl = req.file.path;
 		const title = req.body.title;
 		const content = req.body.content;
+		let creator;
 		const post = new Post({
 						title,
 						content,
 						imageUrl,
-						creator: {
-								name: 'Yan Kairalla'
-						}  
+						creator: req.userId
+						  
 		});
 		post.save()
 		.then(result => {
-				console.log(result);
-				res.status(201).json({
-						message: 'Post created successfully!',
-						post  : result 
-				}); 
+				return User.findById(req.userId);
+		})
+		.then( user => {
+			creator = user;
+			user.posts.push(post);
+			return user.save();
+		})
+		.then(result => {
+			res.status(201).json({
+				message: 'Post created successfully!',
+				post,
+				creator:  { _id: creator._id, name: creator.name }
+			}); 
 		})
 		.catch(err => {
 				if(!err.statusCode) {
@@ -130,6 +139,14 @@ exports.updatePost = (req, res, next) => {
 						error.statusCode = 404;
 						throw error;
 				}
+				if(post.creator.toString() !== req.userId) {
+					console.log(post.creator.toString());
+					console.log(req.userId);
+					const error = new Error('Not authorized!');
+					error.statusCode = 403;
+					throw error;
+				}
+
 				if(imageUrl !== post.imageUrl) {
 					clearImage(post.imageUrl);
 				}	
@@ -158,12 +175,23 @@ exports.deletePost = (req, res, next) => {
 			error.statusCode = 404;
 			throw error;
 		}
+		if(post.creator.toString() !== req.userId) {
+			const error = new Error('Not authorized!');
+			error.statusCode = 403;
+			throw error;
+		}	
 		//check logged in user
 		clearImage(post.imageUrl);
 		return Post.findByIdAndRemove(postId);
 	})
 	.then(result => {
-		console.log(result);
+		return User.findById(req.userId);
+	})
+	.then(user => {
+		user.posts.pull(postId);
+		return user.save();
+	})
+	.then(result => {
 		res.status(200).json({ message: 'Deleted Post.' });
 	})
 	.catch(err => {
